Add files via upload
The purpose of this model is to analyze the customer's data with respect to annual income and spend score. Clustering is to identify the cluster of customers and for each of them, we deploy different marketing strategies and business strategies, with high annual income but a low spending score is our target. We can send them better advertisements to track them more and to track them better so that they can purchase more. this model is trained by using k-means clustering with WCSS concept and #Pandas, #Numpy, #Matplotlib,